GG-28799 Added support of atomic caches.

+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/persistence/checkpoint/CheckpointEntry.java
@@ -206,9 +206,11 @@ public class CheckpointEntry {
         }
 
         /**
+         * Returns number of partitions.
          *
+         * @return Number of partitions.
          */
-        public long size(){
+        public int size(){
             return idx;
         }
